# Improve `ILIKE` contains performance